Refactor removePrefix util function
